requirements.txt
crewai
pypdf
cerebras-cloud-sdk
flask
markdown2 